Refactor: Extend ipython-sql
Removes several functions that can be replaced with direct usage of
ipython-sql in the notebook, and replace all remaining sql calls with
calls to ipython-sql. This means we'll use the same connection
information as has been set for ipython-sql.

Also monkeypatches ipython-sql to add a comment to every sql query (as
the previous code did).

https://github.com/open-contracting/kingfisher-colab/issues/29